[DEPLOY] Make it possible to deploy a 'standalone' version of the dev box for CI/demo purposes
This method allows developers to share their local checkout of the main repo with the dev VM by default, allowing active modifications, or alternatively to deploy the dev environment such that the app is deployed directly to it as in production. The latter method is used for CI for containment and production-mirroring purposes.

Also use rvm, a global ruby install and binary ruby (for speed)