[DEPLOY][BRAIN] Avoid dependency on git-based gem since chef is having a hard time installing it
Seeing failures when running the tests saying that the git repo is not checked out, even though chef did bundle install. A manual bundle install fixes the issue. Has to be an rbenv-related thing. For now, we can side-step the original issue the patched gem was used for in a different way.